+++ Source/EventFlow.MsSql/IMsSqlConfiguration.cs
@@ -21,11 +21,15 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+using EventFlow.Core;
 using EventFlow.Sql.Connections;
 
 namespace EventFlow.MsSql
 {
     public interface IMsSqlConfiguration : ISqlConfiguration<IMsSqlConfiguration>
     {
+        RetryDelay ServerBusyRetryDelay { get; }
+
+        IMsSqlConfiguration SetServerBusyRetryDelay(RetryDelay retryDelay);
     }
 }